rename the main file and add argument parser
